tint-system/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
databases[postgresql]
pydantic>=2.0.0
python-dotenv
aioredis
redis